Extract and restructure hooks
