+++ src/hooks/useForceUpdate.ts
@@ -1,0 +1,7 @@
+import { useState } from 'react';
+
+
+export function useForceUpdate() {
+    const [_, setValue] = useState(0);
+    return () => setValue(value => value + 1);
+}
